requirements.txt
streamlit
plotly
google-generativeai
pyyaml